+++ computer/local_server.py
@@ -75,56 +75,7 @@ class LocalServer(SizelessServer):
         return proc.stdin, proc.stdout, proc.stderr
 
 if __name__ == '__main__':
-    server = LocalServer(1, {})
-    stdin, stdout, stderr = server.run_interactive('python -c "import sys\nprint \'Welcome.  More input!\'\nsys.stdout.flush()\nwhile True:\n  print \'Got: \' + raw_input()\n  sys.stdout.flush()\nprint \'Done.\'"')
+    from interact.main import handle
 
-    # Reading suggestion from http://stackoverflow.com/questions/375427/non-blocking-read-on-a-subprocess-pipe-in-python
-    # Writing suggestion from http://stackoverflow.com/questions/3762881/how-do-i-check-if-stdin-has-some-data
-    import sys, select
-    from threading  import Thread
-
-    try:
-        from Queue import Queue, Empty
-    except ImportError:
-        from queue import Queue, Empty  # python 3.x
-
-    def enqueue_output(out, queue):
-        for line in iter(out.readline, b''):
-            queue.put(line)
-        out.close()
-
-    qout = Queue()
-    tout = Thread(target=enqueue_output, args=(stdout, qout))
-    tout.daemon = True # thread dies with the program
-    tout.start()
-
-    qerr = Queue()
-    terr = Thread(target=enqueue_output, args=(stderr, qerr))
-    terr.daemon = True # thread dies with the program
-    terr.start()
-
-    while True:
-        stdin.flush()
-        sys.stdout.flush()
-
-        try:
-            line = qout.get_nowait()
-        except Empty:
-            pass
-        else:
-            print line
-
-        try:
-            line = qerr.get_nowait()
-        except Empty:
-            pass
-        else:
-            print "ERROR:", line
-
-        hasinputs = select.select([sys.stdin],[],[],0.0)[0]
-        for hasinput in hasinputs:
-            line = sys.stdin.readline()
-            stdin.write(line)
-
-            if line[-1] == '\n':
-                line = line[:-1]
+    server = LocalServer((), 1, {})
+    handle(server)
